requirements.txt
Flask==2.2.2
Pyrebase4==4.6.0
requests==2.28.2
Werkzeug==2.2.2
gunicorn==20.1.0